+++ yagura-v0.35.0/internal/recvcheck/recvcheck.go
@@ -18,6 +18,7 @@
 package recvcheck
 
 import (
+	"errors"
 	"go/ast"
 	"go/parser"
 	"go/scanner"
@@ -127,7 +128,8 @@ func collectMethods(path, src string, byType map[string][]method, r *Report) {
 	f, err := parser.ParseFile(fset, path, src, 0)
 	if err != nil {
 		line := 1
-		if el, ok := err.(scanner.ErrorList); ok && len(el) > 0 {
+		var el scanner.ErrorList
+		if errors.As(err, &el) && len(el) > 0 {
 			line = el[0].Pos.Line
 		}
 		r.Findings = append(r.Findings, Finding{