+++ yagura-v0.35.0/internal/returncheck/returncheck.go
@@ -20,6 +20,7 @@
 package returncheck
 
 import (
+	"errors"
 	"go/ast"
 	"go/parser"
 	"go/scanner"
@@ -112,7 +113,8 @@ func scanFile(path, src string, threshold int) fileScanResult {
 	f, err := parser.ParseFile(fset, path, src, parser.AllErrors)
 	if err != nil {
 		line := 1
-		if el, ok := err.(scanner.ErrorList); ok && len(el) > 0 {
+		var el scanner.ErrorList
+		if errors.As(err, &el) && len(el) > 0 {
 			line = el[0].Pos.Line
 		}
 		res.Findings = append(res.Findings, Finding{